Local branch amd-gfx a73e0939a13e Merged main:c866f8544c92 into amd-gfx:f94846f3674b
Remote branch main 5e9aed1be5a5 [lldb] [test] Mark TestNonStop as LLGS-specific

+++ lld/MachO/SyntheticSections.cpp
@@ -1250,7 +1250,8 @@ void CodeSignatureSection::writeHashes(uint8_t *buf) const {
   uint8_t *hashes = buf + fileOff + allHeadersSize;
   parallelFor(0, getBlockCount(), [&](size_t i) {
     sha256(buf + i * blockSize,
-           std::min(static_cast<size_t>(fileOff - i * blockSize), blockSize),
+           std::min(static_cast<size_t>(fileOff - i * blockSize),
+                    static_cast<size_t>(blockSize)),
            hashes + i * hashSize);
   });
 #if defined(__APPLE__)
